server/requirements.txt
aiocron==1.3
aiohttp==3.5.4
aiohttp-cors==0.7.0
aiohttp-jinja2==1.1.0
aiojobs==0.2.2
APScheduler==3.5.3
args==0.1.0
async-timeout==3.0.1
asyncpg==0.18.3
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
clients==1.0
config==0.4.0
Django==2.1.5
django-scheduler==0.8.8
helpers==0.2.0
icalendar==4.0.3
idna==2.8
Jinja2==2.10
MarkupSafe==1.1.1
multidict==4.5.2
oauth==1.0.1
phonenumbers==8.10.8
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
requests==2.21.0
schedule==0.6.0
six==1.12.0
tzlocal==1.5.1
urllib3==1.24.1
yarl==1.3.0
